@@ -17,32 +17,32 @@
   },
   "private": true,
   "devDependencies": {
-    "@angular-devkit/architect": "^0.1303.11",
-    "@angular-devkit/build-angular": "~13.3.11",
+    "@angular-devkit/architect": "^0.1402.13",
+    "@angular-devkit/build-angular": "^14.2.13",
     "@angular-devkit/build-optimizer": "^0.11.2",
-    "@angular-devkit/core": "^13.3.11",
+    "@angular-devkit/core": "^14.2.13",
     "@angular-eslint/builder": "12.7.0",
     "@angular-eslint/eslint-plugin": "13.5.0",
     "@angular-eslint/eslint-plugin-template": "13.5.0",
     "@angular-eslint/schematics": "14.4.0",
     "@angular-eslint/template-parser": "12.7.0",
-    "@angular/animations": "13.4.0",
+    "@angular/animations": "14.3.0",
     "@angular/cdk": "13.3.9",
-    "@angular/cli": "13.3.11",
-    "@angular/common": "13.4.0",
-    "@angular/compiler": "^13.4.0",
-    "@angular/compiler-cli": "^13.4.0",
-    "@angular/core": "13.4.0",
-    "@angular/elements": "13.4.0",
+    "@angular/cli": "14.2.13",
+    "@angular/common": "14.3.0",
+    "@angular/compiler": "^14.3.0",
+    "@angular/compiler-cli": "^14.3.0",
+    "@angular/core": "14.3.0",
+    "@angular/elements": "14.3.0",
     "@angular/flex-layout": "^13.0.0-beta.38",
-    "@angular/forms": "13.4.0",
-    "@angular/language-service": "13.4.0",
+    "@angular/forms": "14.3.0",
+    "@angular/language-service": "14.3.0",
     "@angular/material": "13.3.9",
     "@angular/material-moment-adapter": "^13.3.9",
-    "@angular/platform-browser": "13.4.0",
-    "@angular/platform-browser-dynamic": "13.4.0",
-    "@angular/router": "13.4.0",
-    "@angular/service-worker": "13.4.0",
+    "@angular/platform-browser": "14.3.0",
+    "@angular/platform-browser-dynamic": "14.3.0",
+    "@angular/router": "14.3.0",
+    "@angular/service-worker": "14.3.0",
     "@kurkle/color": "^0.1.9",
     "@ngtools/webpack": "^13.3.11",
     "@ngxs/logger-plugin": "^3.4.3",
